Separate the LTIGradesSenderControlFile class into its own file
